requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
builtwith==1.3.4
certifi==2020.6.20
chardet==3.0.4
future==0.18.2
idna==2.10
lxml==4.5.1
Pillow==7.2.0
PyMySQL==0.9.3
python-whois==0.7.3
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.9
